Payment, one-to-one association

+++ src/main/java/com/educandoweb/course2/entities/User.java
@@ -23,7 +23,7 @@ public class User  implements Serializable{
     
     @Id //Declarando chave primaria
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
     private String phone;
@@ -38,7 +38,7 @@ public class User  implements Serializable{
     }
 
     //COSNTRUTOR SOBRECARREGADO
-    public User(Integer id, String name, String email, String phone, String password) {
+    public User(Long id, String name, String email, String phone, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -74,11 +74,11 @@ public class User  implements Serializable{
 
     
     //GET AND SET
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
